+++ src/Adapter/ColorScheme/VARIABLE.mjs
@@ -1,3 +1,5 @@
+export const VARIABLE_ACCENT = "accent";
+
 export const VARIABLE_BACKGROUND = "background";
 
 export const VARIABLE_FOREGROUND = "foreground";