+++ CHANGELOG.md
@@ -6,6 +6,13 @@ Changes:
 
 \-
 
+## v2022-10-13-1
+
+Changes:
+
+- Accent color
+- Specific get variable function
+
 ## v2022-10-12-1
 
 Changes: